server/pyproject.toml
[project]
name = "rag-ssw-server"
version = "0.1.0"
description = "RAG-SSW Backend Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "weaviate-client>=4.5.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.0",
    "langchain-openai>=0.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "docling>=2.0.0",
    "docling-datamodel>=2.0.0",
    "pillow>=10.0.0",
    "loguru>=0.7.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
